Experiment/experiment_2/requirements.txt
streamlit
pdfplumber
pytesseract
pdf2image
Pillow
llama-cpp-python
